Implementation of the detached app and of the euphonic widgets:
- standard path
- linear path
- Q planes
